Additional changes to support move from src to lib
