+++ specs/Modal.spec.js
@@ -4,8 +4,8 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import TestUtils from 'react-addons-test-utils';
 import createReactClass from 'create-react-class';
-import Modal from '../lib/components/Modal';
-import * as ariaAppHider from '../lib/helpers/ariaAppHider';
+import Modal from '../src/components/Modal';
+import * as ariaAppHider from '../src/helpers/ariaAppHider';
 import {
   isBodyWithReactModalOpenClass,
   contentAttribute,
