+++ specs/helper.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Modal from '../lib/components/Modal';
+import Modal from '../src/components/Modal';
 import TestUtils from 'react-addons-test-utils';
 
 const divStack = [];
